% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{done_line}
\alias{done_line}
\alias{warning_line}
\alias{info_line}
\title{Print clean cli information text}
\usage{
done_line(..., .envir = parent.frame())

warning_line(..., .envir = parent.frame())

info_line(..., .envir = parent.frame())
}
\arguments{
\item{...}{Any params usable by \code{\link[glue]{glue}}}

\item{.envir}{Environment}
}
\value{
invisible
}
\description{
Print clean cli information text
}
\examples{
done_line("wooo....bullets")
warning_line("wooo....bullets")
info_line("wooo....bullets")
}
