% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{reproducibility_receipt}
\alias{reproducibility_receipt}
\title{Create Reproducibility Receipt}
\usage{
reproducibility_receipt(render = TRUE)
}
\arguments{
\item{render}{Whether to render the github markdown results for preview in the viewer}
}
\value{
Invisible return of content. Content added to clipboard for copy and paste
}
\description{
In our pull request template, the end of the template
provides a section for a "reproducibility receipt". The code
below provides that receipt that can be copy and pasted into all
pull requests.
}
\examples{
reproducibility_receipt()
}
