% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_templates.R
\name{use_aclu_review_template}
\alias{use_aclu_review_template}
\alias{use_aclu_pr_template}
\alias{use_aclu_issue_template}
\alias{use_aclu_label_template}
\title{ACLU Review Template}
\usage{
use_aclu_review_template(type = "analysis", repo = ".")

use_aclu_pr_template(type = "analysis", repo = ".")

use_aclu_issue_template(type = "analysis", repo = ".")

use_aclu_label_template(type = "analysis", repo = ".")
}
\arguments{
\item{type}{Type of repository. Defaults to "analysis" which is the only fully built out template in base version}

\item{repo}{Repository. Defaults to "." - base path. We use `repo` instead of `path` here to align with `git2r` standards.}
}
\value{
invisible. Provides cli based messages as completes
}
\description{
Apply ACLU Analytics team review templates to your github repository
to help standardize reviews and ensure quality and accuracy of our work
}
\examples{
# apply all of the template items
use_aclu_review_template()

# apply the PR template to the repository
use_aclu_pr_template()

# apply issue templates to the repository
use_aclu_issue_template()

# apply labels to the repository
use_aclu_label_template()

}
