% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{warning_line}
\alias{warning_line}
\title{Print clean warning text during project setup}
\usage{
warning_line(..., .envir = parent.frame())
}
\arguments{
\item{...}{Any params usable by \code{\link[glue]{glue}}}

\item{.envir}{Environment}
}
\value{
invisible
}
\description{
Print clean warning text during project setup
}
\examples{
warning_line("wooo....bullets")
}
