% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{done_line}
\alias{done_line}
\title{Print clean done text during project setup}
\usage{
done_line(..., .envir = parent.frame())
}
\arguments{
\item{...}{Any params usable by \code{\link[glue]{glue}}}

\item{.envir}{Environment}
}
\value{
invisible
}
\description{
Print clean done text during project setup
}
\examples{
done_line("wooo....bullets")
}
